Actually let meson use pre-generated introspection file
Unlike autoconf, meson scripts actually aborted if the program 'xxd' was
absent regardless of whether the pre-generated introspection file was
found or not. Make xxd optional, and if it is not found print a warning
and make the dependency object xml_introspection_h point at the relevant
file in the source directory instead of generating a new one in the build
directory. If that file does not exist either, abort.